requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0
allure-pytest==2.13.2
